Separate transformer summarisation
